requirements.txt
# DDS-LLM Orchestrator Requirements

# Core
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0

# DDS (optional)
cyclonedds>=0.10.0

# Server
uvicorn>=0.25.0
fastapi>=0.109.0

# Utilities
requests>=2.31.0

# For development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
